+++ lib/oh_my_flutter.dart
@@ -23,7 +23,7 @@ export 'src/widgets/marquee/marquee.dart' show Marquee;
 export 'src/widgets/marquee/marquee_direction.dart' show MarqueeDirection;
 export 'src/widgets/maybe_safe_area/maybe_safe_area.dart' show MaybeSafeArea;
 export 'src/widgets/morph/morph.dart'
-    show Morph, MorphDescendant, MorphEndpoint, MorphEndpointContext, MorphFlight, MorphFlightDelegate;
+    show Morph, MorphDescendant, MorphEndpoint, MorphEndpointContext, MorphFlight, MorphFlightDelegate, MorphForeground;
 export 'src/widgets/morph/morph_descendant_flight_behavior.dart' show MorphDescendantFlightBehavior;
 export 'src/widgets/morph/morph_flight_kind.dart' show MorphFlightKind;
 export 'src/widgets/motion/motion.dart'